pyproject.toml
[project]
name = "soy-factory"
version = "0.1.0"
description = "Smart Soy Sauce Factory"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "PyQt6>=6.4.0",
    "alembic>=1.14.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "qtawesome>=1.4.1",
    "bcrypt>=5.0.0",
    "pyserial>=3.5",
    "PyQt6-Fluent-Widgets>=1.6.0",
    "httpx>=0.28.1",
]
